feat: support cast decimal to utf8
