+++ crates/polars-core/src/fmt.rs
@@ -1153,125 +1153,15 @@ impl Series {
     }
 }
 
+#[inline]
 #[cfg(feature = "dtype-decimal")]
-mod decimal {
-    use std::fmt::Formatter;
-    use std::{fmt, ptr, str};
+pub fn fmt_decimal(f: &mut Formatter<'_>, v: i128, scale: usize) -> fmt::Result {
+    use arrow::legacy::compute::decimal::format_decimal;
 
-    use crate::fmt::{fmt_float_string, get_trim_decimal_zeros};
-
-    const BUF_LEN: usize = 48;
-
-    #[derive(Clone, Copy)]
-    pub struct FormatBuffer {
-        data: [u8; BUF_LEN],
-        len: usize,
-    }
-
-    impl FormatBuffer {
-        #[inline]
-        pub const fn new() -> Self {
-            Self {
-                data: [0; BUF_LEN],
-                len: 0,
-            }
-        }
-
-        #[inline]
-        pub fn as_str(&self) -> &str {
-            unsafe { str::from_utf8_unchecked(&self.data[..self.len]) }
-        }
-    }
-
-    const POW10: [i128; 38] = [
-        1,
-        10,
-        100,
-        1000,
-        10000,
-        100000,
-        1000000,
-        10000000,
-        100000000,
-        1000000000,
-        10000000000,
-        100000000000,
-        1000000000000,
-        10000000000000,
-        100000000000000,
-        1000000000000000,
-        10000000000000000,
-        100000000000000000,
-        1000000000000000000,
-        10000000000000000000,
-        100000000000000000000,
-        1000000000000000000000,
-        10000000000000000000000,
-        100000000000000000000000,
-        1000000000000000000000000,
-        10000000000000000000000000,
-        100000000000000000000000000,
-        1000000000000000000000000000,
-        10000000000000000000000000000,
-        100000000000000000000000000000,
-        1000000000000000000000000000000,
-        10000000000000000000000000000000,
-        100000000000000000000000000000000,
-        1000000000000000000000000000000000,
-        10000000000000000000000000000000000,
-        100000000000000000000000000000000000,
-        1000000000000000000000000000000000000,
-        10000000000000000000000000000000000000,
-    ];
-
-    pub fn format_decimal(v: i128, scale: usize, trim_zeros: bool) -> FormatBuffer {
-        const ZEROS: [u8; BUF_LEN] = [b'0'; BUF_LEN];
-
-        let mut buf = FormatBuffer::new();
-        let factor = POW10[scale]; //10_i128.pow(scale as _);
-        let (div, rem) = (v / factor, v.abs() % factor);
-
-        unsafe {
-            let mut ptr = buf.data.as_mut_ptr();
-            if div == 0 && v < 0 {
-                *ptr = b'-';
-                ptr = ptr.add(1);
-                buf.len = 1;
-            }
-            let n_whole = itoap::write_to_ptr(ptr, div);
-            buf.len += n_whole;
-            if rem != 0 {
-                ptr = ptr.add(n_whole);
-                *ptr = b'.';
-                ptr = ptr.add(1);
-                let mut frac_buf = [0_u8; BUF_LEN];
-                let n_frac = itoap::write_to_ptr(frac_buf.as_mut_ptr(), rem);
-                ptr::copy_nonoverlapping(ZEROS.as_ptr(), ptr, scale - n_frac);
-                ptr = ptr.add(scale - n_frac);
-                ptr::copy_nonoverlapping(frac_buf.as_mut_ptr(), ptr, n_frac);
-                buf.len += 1 + scale;
-                if trim_zeros {
-                    ptr = ptr.add(n_frac - 1);
-                    while *ptr == b'0' {
-                        ptr = ptr.sub(1);
-                        buf.len -= 1;
-                    }
-                }
-            }
-        }
-
-        buf
-    }
-
-    #[inline]
-    pub fn fmt_decimal(f: &mut Formatter<'_>, v: i128, scale: usize) -> fmt::Result {
-        let trim_zeros = get_trim_decimal_zeros();
-        f.write_str(fmt_float_string(format_decimal(v, scale, trim_zeros).as_str()).as_str())
-    }
+    let trim_zeros = get_trim_decimal_zeros();
+    let repr = format_decimal(v, scale, trim_zeros);
+    f.write_str(fmt_float_string(repr.as_str()).as_str())
 }
-
-#[cfg(feature = "dtype-decimal")]
-pub use decimal::fmt_decimal;
 
 #[cfg(all(
     test,
